Revamped Feed Forward and AddNode Mutation

+++ src/Environment.java
@@ -8,9 +8,9 @@ public interface Environment {
     /**
      * Fitness function corresponding to the environment.
      * @param g The genome to evaluate the fitness of.
-     * @return (float) Fitness of the specific genome
+     * @return (double) Fitness of the specific genome
      */
-    float fitness(Genome g);
+    double fitness(Genome g);
 
 
 }
